requirements.txt
fastapi~=0.110.0
pandas~=2.2.1
plotly~=5.19.0
scikit-learn~=1.4.1.post1
requests~=2.31.0
beautifulsoup4~=4.12.3
pymongo~=4.6.2
uvicorn~=0.28.0
python-dotenv~=1.0.1